Split out target group tests
